update css and fix some bugs

+++ src/views/Admin.vue
@@ -1,30 +1,28 @@
 <template>
-  <div class="admin">
-    <loading :loading="loading"></loading>
-    <div class="box" style="margin: 1vh 8vw;" v-for="u in users">
-      <div class="title is-4" style="margin-left: 10px;">{{ u.name }}
-        <div class="title is-5">{{ u.role }} &nbsp; {{ u.group }}</div>
-        <button class="button is-danger" style="margin: 0 20px;" @click="deleteUser(u)">
-          <span class="icon">
-            <i class="mdi mdi-24px mdi-delete"></i>
-          </span>
-        </button>
-        <v-switch v-model="u.enable" @change="putUser(u)"></v-switch>
-      </div>
+  <div class="admin pt-4">
+    <loading :loading="loading" />
+    <div class="box m-4" v-for="u in users">
+      <h1 class="title is-4 m-0">
+        {{ u.name }}
+        <span class="icon m-1" @click="deleteUser(u)">
+          <i class="mdi mdi-24px mdi-trash-can-outline" style="color: red" />
+        </span>
+      </h1>
+      <p class="is-5">{{ u.role }} &nbsp; <code>{{ u.group }}</code></p>
+      <v-switch v-model="u.enable" @change="putUser(u)"></v-switch>
     </div>
   </div>
 </template>
 
 <script setup>
 import { token } from '../plugins/state.js'
-const opt = { headers: { token: token() } }
 ref: users = []
 ref: loading = true
 getUser()
 
 async function getUser () {
   loading = true
-  await axios.get('/api/user', opt)
+  await axios.get('/api/user', token())
     .then(res => {
       users = res.data
       users.sort((a, b) => b.enable && !a.enable)
@@ -32,6 +30,7 @@ async function getUser () {
     .catch(err => { window.Swal.fire('错误', err.response ? err.response.data : '网络错误', 'error') })
   loading = false
 }
+
 async function putUser (u) {
   loading = true
   await axios.put('/api/user/' + u._id, { enable: u.enable }, opt)
@@ -39,8 +38,19 @@ async function putUser (u) {
     .catch(err => { window.Swal.fire('错误', err.response ? err.response.data : '网络错误', 'error') })
   loading = false
 }
+
 async function deleteUser (u) {
   loading = true
+  const r = await Swal.fire({
+    title: '你确定要删除吗？',
+    text: "此操作不可以撤销",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: '确认',
+    cancelButtonText: '取消'
+  })
+  if (!r.isConfirmed) return
+
   await axios.delete('/api/user/' + u._id, opt)
     .then(res => { window.Swal.fire('成功', res.data, 'success') })
     .catch(err => { window.Swal.fire('错误', err.response ? err.response.data : '网络错误', 'error') })
@@ -51,8 +61,7 @@ async function deleteUser (u) {
 
 <style scoped>
 div.admin {
-  min-height: 100vh;
-  padding-top: 80px;
+  min-height: 93vh;
   background: #eee;
 }
 </style>
